requirements.txt
ultralytics==8.3.27
opencv-python
pillow
pytesseract
fasttext
numpy
